FTS-906:Change int user_filesize to long long

+++ src/db/mysql/MySqlAPI.cpp
@@ -2735,7 +2735,7 @@ std::vector<TransferState> MySqlAPI::getStateOfDeleteInternal(soci::session& sql
             else
                 ret.file_metadata = "";
 
-            ret.user_filesize = it->get<int>("user_filesize", 0);
+            ret.user_filesize = it->get<long long>("user_filesize", 0);
 
             ret.source_se = it->get<std::string>("source_se");
             ret.dest_se = "";
@@ -2817,7 +2817,7 @@ std::vector<TransferState> MySqlAPI::getStateOfTransferInternal(soci::session& s
             ret.vo_name = it->get<std::string>("vo_name");
             ret.job_metadata = it->get<std::string>("job_metadata","");
             ret.retry_max = it->get<int>("retry_max",0);
-            ret.user_filesize = it->get<int>("user_filesize", 0);
+            ret.user_filesize = it->get<long long>("user_filesize", 0);
             ret.file_id = it->get<unsigned long long>("file_id");
             ret.file_state = it->get<std::string>("file_state");
             ret.timestamp = milliseconds_since_epoch();
